+++ src/components/PostContents/index.tsx
@@ -1,6 +1,6 @@
-import { useState } from 'react';
-import MarkdownRender from '@components/MarkdownRender';
-import HeadingAnchor, { markDownID } from '@components/HeadingAnchor';
+import { useState, useEffect, useCallback } from 'react';
+import MarkdownRender from './MarkdownRender';
+import AnchorList, { markDownID } from '@components/PostContents/AnchorList';
 
 interface Props {
   contents: string;
@@ -12,15 +12,20 @@ const PostContents: React.FC<Props> = ({ contents }) => {
 
   const syncAnchorList = (value: markDownID[]) => {
     setAnchorList([...value]);
-  };
-  const syncFocusAnchor = (anchorID: string) => {
-    setFocusAnchor(anchorID);
-  };
+  }
+
+  // const updateFocusAnchor = useCallback((anchorID: string) => {
+  //   setFocusAnchor(anchorID);
+  // }, [focusAnchor]);
+
+  const updateFocusAnchor = (anchorID:string) => {
+    setFocusAnchor(anchorID)
+  }
 
   return (
-    <div className='postcontents' >
-      <HeadingAnchor anchorList={anchorList} focusAnchor={focusAnchor} />
-      <MarkdownRender markdown={contents} updateFocusAnchor ={syncFocusAnchor} updateAnchorList={syncAnchorList} />
+    <div className='postcontents'>
+      <AnchorList anchorList={anchorList} focusAnchor={focusAnchor} />
+      <MarkdownRender markdown={contents} updateFocusAnchor={updateFocusAnchor} updateAnchorList={syncAnchorList} />
     </div>
   );
 };
